LayoutTests/js/arrowfunction-bind-expected.txt

Tests for ES6 arrow function, this should be overwritten during invoke bind

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS d.y().bind(e, 'ley')() is 'barley'
PASS successfullyParsed is true

TEST COMPLETE
